pyproject.toml
[tool.poetry]
name = "django-ecommerce"
version = "0.1.0"
description = "A Django eCommerce platform"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
asgiref = "3.8.1"
Django = "5.2.1"
django-ckeditor = "6.7.2"
django-easy-audit = "1.3.7"
django-jazzmin = "3.0.1"
django-js-asset = "3.1.2"
django-shortuuidfield = "0.1.3"
django-taggit = "6.1.0"
Pillow = "11.2.1"
python-decouple = "3.8"
shortuuid = "1.0.13"
six = "1.17.0"
sqlparse = "0.5.3"
typing_extensions = "4.13.2"
tzdata = "2025.2"
django-allauth = {extras = ["socialaccount"], version = "^65.8.0"}
django-countries = "^7.6.1"
celery = "^5.5.2"
redis = "4.6.0"
django-celery-results = "^2.6.0" 
django-celery-beat = "^2.8.1"   
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
daphne = "^4.2.0"

[tool.poetry.dev-dependencies]
black = "^23.12.0"
flake8 = "^6.1.0"
pytest = "^7.4.3"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"